Fetch can-merge info when updating a pull-request
Instead of synchronously checking via the GitHub API if a pull request
is mergeable, we can already get the necessary information during Github
event processing so the canMerge() check operates on cached data similar
to how this is done for the Gerrit driver already.

With this we can also remove the additional API requests for the commit
status and check runs since this info can be retrieved via the same GraphQL
request.

Change-Id: I6b4848dcb5d27071deeb6a59642c0e3c03a799da