Update patch set 7
Patch Set 7:

I think it might be cool if users could supply an icon by themselves to be shown in the artifacts tab. But I don't think that we should rely on this. So I'd propose to combine this with either the current or a wildcard approach. In case an icon is supplied we use the icon, otherwise we provide our own based on the artifact type.

I was also thinking about something like a mimetype, but I think mimetypes are too generic for this purpose as we don't want to show the same icon for all html pages or for all text files. E.g. a preview site should get another icon than the test results whereby both would be html pages.

Patch-set: 7