Update patch set 1
Patch Set 1:

This is an alternative to I01a8de81a6a270e666009b8be27392bd44ca8c9d using a central thread pool approach.

Based on this we could also remove the thread pool of the github connection and use the central one instead in a later change.

Patch-set: 1