javascript: publish artifact tarball
The javascript publishing has been broken for some time (first
depends-on).

However, when it was working, this job will extract the javascript
tarball artifact into the destination directory.  This has at least
two problems; it is extracting into the same directory as the main
build .tar.gz/.whl files (so you can't see them), and it also has no
sense of the stable branch which will just overwrite depending on
whatever ran last.

It doesn't feel like we need to extract the site any more.  This job
should just copy the javascript artifact tarball and publish that.

Depends-On: https://review.opendev.org/754230
Depends-On: https://review.opendev.org/754241
Change-Id: I6ef8344fa79f33ae1d8dd59c7db307bd45f1c472