Merge Zookeeper connection methods
Zookeeper connection is done using the "connect_zookeeper"
helper function and using "ZooKeeper#connect" method.

This change merges both into one to consolidate usage.

Additionally a "ZookeeperConnection" is added to allow
save disconnecting using a "with" block.

Last, a way to override Zookeeper properties in test
is hinted here. This will be used in follow-up changes.

Change-Id: I5fa4ddab5f85c658291f1262ee0392a60086846e