Configure redux for development
This enables the redux developer tools for the browser. To make use of
this, one must also install the Redux DevTools extension which is
available for various browsers. The extension visualize all state
transitions in the redux store and also allows changing them manually to
see the effects.

Additionally, it makes use of the "redux-immutable-state-invariant"
package, which throws exception in development mode whenever a state is
mutated directly within an action or reducer.

Unfortunately, I didn't manage to keep the tests green as there seem to
be some weird state mutations somewhere deep within the zuul-web JS code
which I couldn't track down yet. So, redux-immutable-state-invariant is
now only used during development.

Change-Id: I8a8588cd7f5f1b17b247d9700a492e5c1e27f040