Update patch set 2
Patch Set 2: Code-Review-1

I'm still learning all this, so I don't have a reason why, but ...

Something I noticed was on the status page with this, the refresh totally blanks the page.  So like currently you see the refresh spinner go, but the page contents stay the same.  With this, the page goes completely blank, then the spinner goes for about a second, then all the change boxes pop up.

It makes it impossible to use the status page because it keeps disappearing on you every few seconds

Patch-set: 2
Label: Code-Review=-1