Update patch set 2
Patch Set 2:

> I'm still learning all this, so I don't have a reason why, but ...
 > 
 > Something I noticed was on the status page with this, the refresh
 > totally blanks the page.  So like currently you see the refresh
 > spinner go, but the page contents stay the same.  With this, the
 > page goes completely blank, then the spinner goes for about a
 > second, then all the change boxes pop up.
 > 
 > It makes it impossible to use the status page because it keeps
 > disappearing on you every few seconds

Thanks for the hint, ianw. To be honest I didn't notice that. That might be the reason why the old state was set as new state in the status reducer. But AFAIK this doesn't make any sense as the "status" part of the state shouldn't change at all in between. I will investigate this. Maybe there is more to do in the status actions/reducers.

Patch-set: 2
Label: Verified=0