Update patch set 2
Patch Set 2:

> Release note should help, agreed, but is this behavior change one
 > we want to consider as a regression? Is it fixable to be able to
 > retain the old behavior?

I agree that this is a regression we overlooked. It is fixable by adding a higher complexity to the graphql queries (checking if check run permission is there and dynamically alter the graphql query).

We should decide if we want to treat this as a regression and fix it or if we want to require check run permissions in the future.

Nevertheless I don't think we should do another pre 4.0 release due to this given the problems we had the last time doing that.

Patch-set: 2
Reviewer: Gerrit User 16068 <16068@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0