Run jobs automatically on playbook changes
This adds all playbooks for a job, explicit and inherited, to a list
that is checked in the files matcher.  This means that jobs that have
restriced their run to specific files with "files:" will still run
when the playbooks they use are modified.

This is similar in spirit to 7fba932e5dfbabe80b9a1faadf0b0927f7ad01ec
which runs jobs when their config changes.  While you can of course
add playbook files explicitly, you can see plenty of examples where it
is forgotten.  You also need to manually follow the inheritance chain
and parent playbooks, which is error-prone and difficult to maintain.

Change-Id: I6f0715b60bfc80e465ba87a73d278abe746747d9