Update patch set 17
Patch Set 17: Code-Review-1

(1 comment)

Thanks, this generally looks good!

Just a couple of things, a nit on wording inline.

There's something weird going on with the "x" checkbox next to the typeahead?  It always appears, and a lot of the clicking on it does nothing?

When I look at the example

https://www.patternfly.org/v4/documentation/react/components/select/typeahead

it operates how I would expect with the clear button only appearing when you type, and then clicking it clears the field back to the default.  Any ideas what's up?

Patch-set: 17
Label: Code-Review=-1