Increase merge retries and delays in between
GitHub occasionally denies merges because some internal background
process didn't finish necessary work for it. This happens regularly in
our deployment. Currently zuul retries the merge after two seconds and
then fails and triggers a gate reset. However we see frequent
occasions where another retry or longer wait times would have made it
to succeed. Further we also observed that GitHub sometimes lied about
the result of a merge operation by doing the merge successfully but
responding with an error.

In order to handle those issues more robust retry more often and sleep
slightly longer between the retries. Further after getting an error we
still need to check if the PR is merged in case it succeeded 'in the
background'.

This can have a severe impact on the run_handler loop performance so
this should only be combined with async reporting.

Change-Id: I45a08f32a243d0a0bd1492fa0d244f81e351772a