Update patch set 20
Patch Set 20: Code-Review-1

Thanks; I had a closer look at what's going on with the typeahead -- I think it is because we should be setting the current/default value as the placeholder text, and then making sure the selection is "null" when nothing is selected.  That ensures it doesn't show the "clear the box" X check mark that does nothing.

I'd suggest moving the selection into state, and then setting the placeholder text from the incoming prop that gives the currently set location.

It's easier to see in code :), so i've proposed 

https://review.opendev.org/756164 

which I think shows the typeahead with the placeholder text/clear working as in the patternfly examples.

I don't know if that is the best solution, but I'd like to see something like that so we are consistent with patternfly behaviour.

Patch-set: 20
Label: Code-Review=-1