Update patch set 1
Patch Set 1: Code-Review+2

This looks like a good table conversion, thanks.

Putting a title on the page and an explanation of what's being shown probably wouldn't be a bad enhancement.

We've also been putting numbers in labels, which I think could be done here if we wanted as well.

To me it makes all the repeated words stand out even more than it was before.  I wonder if anyone has any ideas on more of an overhaul; maybe putting each tenant in a panel with a treeview navigation or something?

Patch-set: 1
Label: Code-Review=+2