Update patch set 4
Patch Set 4: Code-Review-1

Something weird seems to be up for me, using Firefox 81.0.1

I pulled this change then ran "npm update" to get the new packages, then started the dev server.

When I bring up the page, it starts to load, then the spinner takes over ("fetching") and takes a *very* long time to show the tenants; like 2 minutes long.  Here is a profile screen shot

https://imgur.com/a/Kbd3DvM

in that time, the "this page is running a script that is non-responsive" dialog pops up a few times.

If I choose to stop the script then, then it seems that the tenants page pops up.  However, I get the following in the console

http://paste.openstack.org/show/798721/ which references many times

 trackProperties@http://localhost:3000/static/js/0.chunk.js:326699:31

I assume it's deep inside react, so I have no idea if it helps, but line 326699 is

http://paste.openstack.org/show/798720/

It doesn't seem to make any difference if I have the redux devtools extension installed or not; same thing happens.  I also have the react devtools, I thought the profiling might help, but that is telling me that the react version is too old for profiling?

This is probably something I'm doing wrong :)  But what that is, I have no idea at this point

Patch-set: 4
Label: Code-Review=-1