Update patch set 4
Patch Set 4:

ianw: I'm glad that you checked out the code to try it yourself :)

One note: Since we are using yarn for package management and locking exact versions, you should use "yarn install" to install the latest dependencies specified in the yarn.lock file. I'm not sure if "npm update" takes the lockfile into account - so it might also update some other dependencies which have never versions by now.

I'm not sure about the Firefox part you mentioned first, but the exception you see in Chrome ("Maximum Call Stack Exceeded") is the one that is raised by redux-immutable-state-invariant - just in this case it's not as helpful as the other exceptions it raises in the zuul-web frontend. I managed to solve a few other state mutations in the builds and buildsets actions with the help of redux-immutable-state-invariant (see the child commit that deals with the state mutations in the build reducers).

Usually redux-immutable-state-invariant raises an exception whenever a state mutations is detected within a dispatch (action/reducer) or somewhere ("anywhere" in the related code that deals with those actions. The one you got with the Call Stack Exceeded is exactly the one I wasn't able to solve (see commit message) and why I deactivated redux-immutable-state-invariant for the tests - so it only runs in development mode. It might also be a bug in redux-immutable-state-invariant but I assume this particular exception comes up due to the nature of how it tries to track down state mutations - by copying and comparing a lot of (nested) objects. That's also the reason why it should only run in development mode and not in production. I also found the linked issue while trying to fix this exception but that didn't help either.

What we could do here is that we ignore this particular state mutation (like ignoring a linter rule for a specific line in code) as I currently have no idea how to solve this and it doesn't seem to harm the result (what is shown in the end). But for the remaining work with redux and zuul-web I would really like to keep this extension as it's really helpful when dealing with redux.

Now for the Redux DevTools extension: The Redux DevTools extension is something completely different and shouldn't be related to any exception shown. When opening the developer tools in your browser there should be a new section called "redux" (next to console, network, ...) which shows the state transitions done in the code in a very nice and helpful way.

Patch-set: 4
Reviewer: Gerrit User 27952 <27952@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0