Lock nodes in executor instead of scheduler
To prevent mass-deletion of nodes in case scheduler
looses connection to zookeeper.

The node request deletion is postponed after the
executor first gives a sign of a running job
(by sending data back to scheduler).

The node lock is done exclusively in the executor server.

This change changes the workflow of how the node requests
and node locking works.

Change-Id: Ic70f80e43b40f0785344eddc576b103165db5dda