JS: Don't run tests in watch mode
By default the react-scripts-test command (which is invoked by `yarn
test`) runs in watch mode, meaning it runs the tests on every code
change while intelligently trying to run only the tests for the files
that change.

This may be a matter of personal taste, but I don't quite get a benefit
out of this watch mode.

First of all, running the tests takes quite some time. Even so we only
have a few tests, running those takes around 10 to 20 seconds. If
watching mode is enabled, the tests run whenever a file is changed and
saved. IMHO this is completely unnecessary as I don't want to run the
tests on each file change, but when I think that a change is complete
and fully functional. Dependending on the size of the actual change
using the watch mode spoils the terminal with annoying test failures
as - of course - the change is not complete yet and variables might be
undefined, the syntax could be broken or similar.

Second, running the tests seems quite resource-hungry and it sometimes
slows down the terminal.

Third, I'm not sure if the "intelligent" test selection always works,
so IMHO it's better to manually run all tests (or manually filter for
the relevant tests) when I think it's time to run the tests.

Change-Id: Iea1276fc405b360f1cfcba66cd3f6735cb7c5565