Make Zookeeper mandatory for Scheduler
Scheduler without Zookeeper makes less sense
with the path Scheduler is now going. Currently
there is only one usecase for that in the
Client#validate.

Making Zookeeper a mandatory parameter in
constructor helps to avoid race conditions
or missing reference.

There are 2 places which need to be adapted:

First, is Scheduler#_checkTenantSourceConf:
This method has currently 2 problems. It
is a protected method while it is used
also outside its class. And it takes a
ConfigurationRegistry argument which
it ignores and uses the instance's
ConnectionRegistry. Since all calls
withing Scheduler class send the
parameter anyway, we can use the parameter
instead of the instance property. Doing
that allows this method to become
a @classmethod and easily used in
Client#validate without instantiating
Scheduler for such call.

Second, TenantParser#getSchema method
which needs only ConnectionRegistry
for it to function properly not the
whole Scheduler isntance. With a small
effort this method can be passed
ConnectionRegistry as a parameter,
become also a @classmethod and
be used in Client#validate without
instantiating Scheduler for such call
either.

The need of instantiating Scheduler
without Zookeeper connection becomes
then obsolete.

Change-Id: I87f652d4c723d4f745b0120ff41905bc120d1846