Delay getting many files to mergers
In the past GitHub only returned at max 300 changed files of a
PR. Therefore zuul has a mechanism to delay getting the file list
using the mergers. Latest GitHub Enterprise seems to have removed this
limitation completely. This leads to large delays in the event
processing if there are incoming events for PRs with many files
(e.g. requesting files for a PR with 30000 files means requesting 300
pages of changed files and leads to a roughly 5 minute delay).

In order to prevent zuul limit this to at max 10 pages of changed
files and if there are more delay to the mergers.

Change-Id: I910ec545c4375103a5ffe5da106c7f5648aa5c72