Instantiate executor client, merger, nodepool and app within Scheduler
Executor client, merger, nodepool and app were instantiated outside the
scheduler and then set using "setX" methods.

Those three components are considered as mandatory and should therefore
be part of all scheduler instances.

This was useful for layout validation where the scheduler was not run
but just instantiated. Since the layout validation does not need to
instantiate a scheduler anymore, this can be simplified by instantiating
these components within the scheduler's constructor.

Change-Id: Ide96a85d17820e3950704577ca6fd0d082e26182