Avoid ref parsing when creating heads
In case a branch name contained the '@' character this was interpreted
by Gitpython according to the format supported by git-rev-parse when
resolving the ref name.

Since we already have the commit for the ref we can use it directly,
that way avoiding any issues caused by parsing the ref name.

Change-Id: I49665c62389245f937317e70f093d33a4bf759d3