Update patch set 12
Patch Set 12: Code-Review-1

I like it :)

One weirdness I'm seeing is that when you pull up the build page, it now shows the "the build provides no logs" page, and then after a few seconds it switches to the "fetching" spinner, and gets the logs.  I think that's a bit confusing.

Patch-set: 12
Label: Code-Review=-1