Cache sql connection names in zuul-web
Currently we ask the scheduler for the sql connection name of a tenant
each time we get an api request that fetches data from the
database. Those requests are processed single threaded by the
scheduler which becomes a severe bottleneck in terms of response
times. Fix this by caching the connection names per tenant.

Change-Id: Id7d2ee9b77b45ade7166b508b7b866762a5b378b