Update patch set 12
Patch Set 12:

ianw: Interesting found. This seems to occur in the time between fetching the build and fetching the manifest/output. During this time, those information is not present yet but it's also not fetching. Thus, https://opendev.org/zuul/zuul/src/branch/master/web/src/pages/Build.jsx#L143 evaluates to false and the EmptyState is shown. This applies for both the task summary (output) and the logs tab (manifest).

Interestingly this pause is much longer in firefox than in chrome. In chrome I can only reproduce it by pausing the state transitions with the redux dev tools.

I have two solutions in mind: I think we either have to store the manifest and the output separately from the build (and combine all three in mapStateToProps() https://opendev.org/zuul/zuul/src/branch/master/web/src/pages/Build.jsx#L274) or set all the fetching values to true in the initialState. I'd rather go for the first solution as this sounds more appropriate. The second one would be more of a hack

Patch-set: 12
Label: Verified=0