UI: Don't show "Invalid Date" for SKIPPED builds in builds table
For SKIPPED builds, the start_time value is always set to null. This
makes the formatting via moment.js fail, resulting in an "Invalid Date"
string being shown for each SKIPPED build in the table.

Since we don't have any start_time information to show for those builds,
just leave the column empty in case the value is not set.

Change-Id: I3537601dcb72a03c07c0e27630d924a61f843557