WIP: ZooKeeper Work abstraction as a replacement for Gearman Jobs
This change add a Work Abstraction to replace Gearman Jobs
in order to replace Gearman with ZooKeeper in future planned
changes.

This change does not affect the runtime itself and is meant
to split the followup changes to improve review process.
It will remain "WIP" until the followup changes needing
this (e.g. https://review.opendev.org/c/744416)
are stable.

Change-Id: I1f30e2283e4a5f20fd90f7d2b6fb1659c60e4ef6