Store secrets keys and SSH keys in Zookeeper
As a preparation for the HA scheduler project secrets keys and SSH keys
will now also be stored in Zookeeper. All private data in Zookeeper will
be encrypted at rest.

Existing keys on the filesystem will be automatically imported into
Zookeeper and new keys will still be available as files for easy backup.

Change-Id: I2a7d1a555f1db1f2178d3bb2f06756ecc8bc7a81