Implementation of Zookeeper backed event queues
Event queues are namespaced by tenant and event type (management, result
and trigger). Events will be stored as sequential nodes in Zookeeper
using the follwing base path:

    /zuul/events/tenants/<tenant-name>/{mangement,result,trigger}/

This allows us to use a simple per-tenant lock in a multi scheduler
deployment that needs to be held while processing events and pipelines.

Change-Id: I9c4835e6a6ac5a83d9f63af146c8dddea9789108