Configure json-server as mock API for development
This configures json-server to run as mock API in development mode.
Using a mock API allows for faster development without having to rely
on a real API. As we already have the possibility to switch between
different APIs, this provides the advantage of another API that
provides the same results every time.

This will allow us to develop and test new UI components or features
without relying on the backend providing the required functionality
(e.g. new API endpoints). In the current state this is often very
difficult, as we would have to wait for the backend changes to be
merged and deployed to make the UI preview build work.

With a mock API we could define a new job which serves the Zuul
dashboard running against this mock API rather than a live Zuul
instance.

When starting the development server via
  $ yarn start
the json-server is started on port 3001 together with the react
development server on port 3000.
By now json-server should provide the following endpoints:
  - http://localhost:3001 -> json-server landing page
  - http://localhost:3001/api/info
  - http://localhost:3001/api/tenants
  - http://localhost:3001/api/config-errors
  - http://localhost:3001/api/builds
  - http://localhost:3001/api/build/<uuid>
  - http://localhost:3001/api/buildsets
  - http://localhost:3001/api/buildset/<uuid>

To make the Zuul dashboard work with the json-server instance, we set
the REACT_APP_ZUUL_API envvar accordingly.

The data provided by json-server is defined in web/tools/db.json which
is automatically generated during startup based on the data in
web/tools/mockData.js. By defining our mock data directly in JavaScript
all developers could use the same data for development and we could even
use this data in our unit tests later on.

By now we only configure a minimal set of data to make the Zuul
dashboard start with the mock API. This includes the central info data,
a single "opendev" tenant, an empty list of config-errors and a bunch of
builds and buildsets. Everything is taken from zuul.opendendev.org.
When requried we could simply extend the data with whatever we need
(projects, build(set)s with different results, nodes, ...).

Change-Id: Ie6ee29cc25ef2e824d06cef25b953d955ca0c2fe