Configure json-server as mock API for development
This configures json-server to run as mock API in development mode.
Using a mock API allows for faster development without having to rely
on a real API. As we already have the possibility to switch between
different APIs, this provides the advantage of another API, that
provides the same results every time.

This will allow us to develop and test new UI components or features
without relying on the backend providing the required functionality
(e.g. new API endpoints). In the current state this is often very
difficult, as we would have to wait for the backend changes to be
merged and deployed to make the UI preview build work.

With a mock API we could define a new job which serves the Zuul
dashboard running against this mock API rather than a live Zuul
instance.

Thus, when starting the development server via
  $ yarn start
we start json-server on port 3001 and configure the react app to use
it as API by setting the REACT_APP_ZUUL_API env var accordingly.

json-server serves the resources defined in the web/tools/db.json file
which is automatically generated during startup based on the data in
web/tools/mockData.js. By defining our mock data directly in JavaScript
all could use the same data for development and we could even use this
data for our unit tests later on.

By now we only configure a bunch of builds and buildsets taken from
zuul.opendev.org to get a start, but we can extend this data whenever
required.

Change-Id: Ie6ee29cc25ef2e824d06cef25b953d955ca0c2fe