Implementation of Zookeeper event watcher
With the event queues in Zookeeper we need a way to set the scheduler's
wake event in case there are new events available. Since queues are
namespaced by tenant and queue type (management, result, trigger) it's
easier to have a central watcher for all the event queues than to
implement the watch on a per-queue basis.

Change-Id: I4dfd2bbee1fc9b8f738c4795cdc0e86e6bf95ce5