WIP UI: Remove stretchable link from builds and buildset table
I tought it would be a good idea to make the whole table row clickable.
However, this made it impossible to mark & copy any text from those
rows.

To avoid the usage of a so called "action columns" (an additional column
containing action buttons) at the end of each row, I added a magnifying
glass icon to the start of each row that contains the link to the build
result page. To reduce visual noise, the icon only shows when hovering
over the table row. Hovering over the icon itself shows a decent growing
effect indicating that this icon is clickable.

Change-Id: I276ff316b17aee932f2724f86e8f0eb1abe73c60