UI: Remove stretchable link from build and buildset table
Making the whole row table row clickable in these pages makes makes it
impossible to mark & copy any text from those rows.  Users have
expressed this is an annoyance worth fixing.

This change removes the stretchable link from the whole table row and
instead allows clicking on the job name or status icon to go to
details about the job.

We have added an optional link prop to title-with-icon; note this is
used in other contexts where it does not make sense to be a link thus
the optional implementation steps.  It is intentional that these links
retain their underline for discoverability.

For the mobile version (which renders the table as a nice card list), we
fall back to the "action column" solution which will be rendered nicely
as a button in the top right corner of each card. For larger screens we
hide this action column.

Note this change is a squash of several other changes to make a
consolidated review; for original details see:

 I788ea37b2ca1899d90160cc52761db0a77f78508
 Ifc54edf9ca348ada8e3bfd07fd135d22f9f08489
 Ifc54edf9ca348ada8e3bfd07fd135d22f9f08489
 Ib4c3a2159e7bad58fcc27eb3ba4b62926ce73099

Co-Authored-By: Ian Wienand <iwienand@redhat.com>
Change-Id: I276ff316b17aee932f2724f86e8f0eb1abe73c60