UI: Remove stretchable link from build and buildset table
I tought it would be a good idea to make the whole table row clickable.
However, this made it impossible to mark & copy any text from those
rows.

This change removes the stretchable link from the whole table row and
instead adds a magnifying glass icon to the start of each row which
contains the link to the build result page. To reduce visual noise, the
icon only shows up when hovering over the table row. Hovering over the
icon itself shows a decent growing effect indicating that this icon is
clickable.

For the mobile version (which renders the table as a nice card list), we
fall back to the "action column" solution which will be rendered nicely
as a button in the top right corner of each card. For larger screens we
hide this action column.

Change-Id: I276ff316b17aee932f2724f86e8f0eb1abe73c60