Update patch set 2
Patch Set 2:

ianw: I've added the missing json-server landing page and updated the commit message with the endpoints that should be served by json-server with that change.

Regarding the missing information in the command line about json-server: The message that json-server is listening on port 3001 is visible for a short time before the development server starts via "create-react-app start". I looks like the "create-react-app start" command removes the previous output from the console to provide this "running in place" feature (not sure how to call that).

But now we get a nice landing page, so everybody can see that json-server is running by visiting http://localhost:3001 ;-)

Patch-set: 2
Label: Verified=0