Update patch set 11
Patch Set 11:

For the record, I'm of the opinion at this point we should wrap on the HTML view of the logs and keep the view fixed width, but provide easy ways to view the logs directly in the browser, which applies no formatting (or i guess the browser formatting of text files).  I have checked the behaviour on github, gitlab and travisci and all do this by default.

This change doesn't exactly work for me.  The preference looks "on" when it loads, but doesn't turn the wrap on.  if i toggle it, it appears to, but then it seems the column for line-numbers isn't wide enough or something:

https://imgur.com/a/UVZq0t9

Patch-set: 11
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0