Improve Elasticsearch reporter doc and driver, changed index name
Some things were not clear enough in Elasticsearch reporter
document.
The Elasticsearch index should include month, because it will
be easier for maintainers to clear the index in the future.
Also remove serialize encoder for the reporter and added
unit tests that is checking if secret is added into the content
which is sent to Elasticsearch.

Change-Id: I80f0bb948856acb3ae8bfcac1d51a34b81e2fb28