Only report dequeue if we have reported start
The dequeue reporting has initially been introduced in order to make
it possible to mark pending check runs as canceled. This is currently
done unconditionally even if zuul hasn't reported the start
already. This leads to occssional spam of canceled check runs that
weren't supposed to be reported at all. For instance we've seen this
on a repo similar to zuul-jobs that is part of all tenants but is only
gated in one tenant. When a PR in such a repo is approved it enters
the gate in all tenants but doesn't run any jobs in all but one
tenant. If the item gets dequeued before the job freezing has been
finished zuul reports canceled check runs from wrong tenants. This
doesn't harm the workflow but leads to user confusion.

A similar problem can be observed when a user creates a PR against a
non-protected branch which typically runs no jobs. In this case an
abandon of the PR can also lead to canceled check run reporting where
zuul was not supposed to report anything at all on the pr.

This can be fixed by skipping dequeue reporting if start hasn't been
reported yet.

Change-Id: Ibd1d8047168dcb5035c90fa25a629f4a7714c0f7