UI: Disable links on skipped builds in build table
Those builds don't have an uuid and thus their link on the build result
page always results in a "Build not found" page. Since we don't have any
result to show, we simply disable the link by giving the magnifying
glass icon.

a disabled look and provide a tooltip with a proper message.

Change-Id: I70dcbc31047fc1c082c5361ad0e3255845902577