web: make build/buildset table details more clickable
This is a small follow-on to I276ff316b17aee932f2724f86e8f0eb1abe73c60
to make the job name/result clickable in the build/buildset view to
take you to the details page.  I think this makes it slightly more
discoverable; I'm a little worried users not overally familiar with
the build details page might miss that it is even there from this view
(I've been pointing some new users at this build overview page lately,
as a way of seeing how their jobs are going).

Change-Id: I788ea37b2ca1899d90160cc52761db0a77f78508