Update patch set 4
Patch Set 4:

Reason was designed for React by the creator of React, and it is based on the same framework. I think the benefits it provides outweigh its learning curve, and perhaps this initial demo is not appropriate and i can show how simpler component could be implemented.

Note that this is replacing the react-ansi component which I find difficult to read and contribute: https://github.com/RaoHai/react-ansi

Patch-set: 4
Label: Verified=0